Attempt to get consistent eval keys across modes
Rather than putting the commands in a prefix map, just use a function to
bind the commands in a specified mode. This makes it easy to keep using
the mostly-standard keys (e.g. C-M-x to evaluate a definition) but
smooth out the differences between modes.

Use a `defvar' to store the mapping between the "thing to be evaluated"
and the key, so that a single change can affect all modes.